Alpha: Send a query (currently not parsed from the frontend representation) to sesame
A query gets send to sesame, the result and the query will be saved to disk. The filename
is a sha1 of the response (ending .txt) or in case of the query the sha1hash_query.txt.
Both filenames are then send back to the client via Ajax and the user may download/open
the generated files in a new window.

All files are saved in /responses/<filenames>